moved opcodes out of insert.c to their own file
