+++ H/goto_ops.h
@@ -1,0 +1,61 @@
+/*
+  goto_ops.ch
+
+  Copyright (C) 2025 Victor Lazzarini
+
+  This file is part of Csound.
+
+  The Csound Library is free software; you can redistribute it
+  and/or modify it under the terms of the GNU Lesser General Public
+  License as published by the Free Software Foundation; either
+  version 2.1 of the License, or (at your option) any later version.
+
+  Csound is distributed in the hope that it will be useful,
+  but WITHOUT ANY WARRANTY; without even the implied warranty of
+  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+  GNU Lesser General Public License for more details.
+
+  You should have received a copy of the GNU Lesser General Public
+  License along with Csound; if not, write to the Free Software
+  Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
+  02110-1301 USA
+*/
+
+#ifndef GOTOPS_H
+#define GOTOPS_H
+
+
+typedef struct {                       
+    OPDS    h;
+    LBLBLK  *lblblk;
+} GOTO;
+
+typedef struct {
+    OPDS    h;
+    int32_t     *cond;
+    LBLBLK  *lblblk;
+} CGOTO;
+
+typedef struct {
+    OPDS    h;
+    MYFLT   *ndxvar, *incr, *limit;
+    LBLBLK  *l;
+} LOOP_OPS;
+
+typedef struct {
+    OPDS    h;
+    MYFLT   *idel, *idur;
+    LBLBLK  *lblblk;
+    int32   cnt1, cnt2;
+} TIMOUT;
+
+typedef struct {
+    OPDS    h;
+} LINK;
+
+typedef struct {
+    OPDS    h;
+    MYFLT   *kInsNo, *kFlags, *kRelease;
+} TURNOFF2;
+
+#endif
